Created hardStateResetAndClearLocalStorage.. ..to sum up all state and localStorage changes when the users logs out or when the user is logged out consequentially after sending an unauthorized request.
