+++ frontend/src/hooks/useLogout.js
@@ -1,9 +1,10 @@
-import { useDispatch } from "react-redux";
 import { useFlashMessage } from "./useFlashMessage";
+import { useHardStateResetAndClearLocalStorage } from "./useHardStateResetAndClearLocalStorage";
 
 export const useLogout = () => {
-  const dispatch = useDispatch();
   const flashMessage = useFlashMessage();
+  const { hardStateResetAndClearLocalStorage } =
+    useHardStateResetAndClearLocalStorage();
 
   const logout = async () => {
     const response = await fetch(
@@ -12,28 +13,13 @@ export const useLogout = () => {
         method: "POST",
         headers: { "Content-Type": "application/json" },
         body: JSON.stringify({ logout: true }),
-        credentials: "include"
+        credentials: "include",
       }
     );
-    if(!response.ok) {
+    if (!response.ok) {
       return flashMessage("ERROR", "Could not log out");
     }
-    if (localStorage.getItem("user")) {
-      localStorage.removeItem("user");
-    }
-    if (localStorage.getItem("newImg")) {
-      localStorage.removeItem("newImg");
-    }
-    if (localStorage.getItem("username")) {
-      localStorage.removeItem("username");
-    }
-    dispatch({ type: "LOGOUT" });
-    dispatch({ type: "RESET_WORKOUTS_STATE" });
-    dispatch({ type: "RESET_ROUTINE_BALANCE_STATE"});
-    dispatch({ type: "RESET_COMPONENTS_STATE" });
-    dispatch({ type: "RESET_PAGE_STATE" });
-    dispatch({ type: "RESET_QUERY_STATE"});
-    return
+    return hardStateResetAndClearLocalStorage();
   };
   return { logout };
 };
